added to model, service and dao for point count system
