+++ src/main/java/evg/testt/dao/UserPointsDao.java
@@ -1,0 +1,7 @@
+package evg.testt.dao;
+
+import evg.testt.model.UserPoints;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserPointsDao  extends JpaRepository<UserPoints, Integer>{
+}
