+++ src/main/java/evg/testt/model/UserData.java
@@ -16,27 +16,6 @@ public class UserData extends BaseModel{
     @JoinColumn(name = "user_id")
     private User user;
 
-    private Integer knowledgeLevelPoints;
-
-    private Integer authorshipPoints;
-
-
-    public Integer getKnowledgeLevelPoints() {
-        return knowledgeLevelPoints;
-    }
-
-    public void setKnowledgeLevelPoints(Integer knowledgeLevelPoints) {
-        this.knowledgeLevelPoints = knowledgeLevelPoints;
-    }
-
-    public Integer getAuthorshipPoints() {
-        return authorshipPoints;
-    }
-
-    public void setAuthorshipPoints(Integer authorshipPoints) {
-        this.authorshipPoints = authorshipPoints;
-    }
-
     public User getUser() {
         return user;
     }
@@ -68,22 +47,6 @@ public class UserData extends BaseModel{
     public void setAge(Integer age) {
         this.age = age;
     }
-
-
-    public void increaseKnowledgeLevel(Integer increment) {
-        knowledgeLevelPoints += increment;
-    }
-    public void decreaseKnowledgeLevel(Integer decrement) {
-        knowledgeLevelPoints -= decrement;
-    }
-
-    public void increaseAuthorshipPoints(Integer increment) {
-        authorshipPoints += increment;
-    }
-    public void decreaseAuthorshipPoints(Integer decrement) {
-        authorshipPoints -= decrement;
-    }
-
 
     public static Builder newBuilder() {
         return new UserData().new Builder();
